Safari-609.1.11.1/LayoutTests/js/dfg-arguments-mutated-structure-expected.txt

Tests effect of mutating Arguments object's structure.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo() is true on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
